split out TemporaryVectorStore
