+++ searchlib/src/vespa/searchlib/tensor/angular_distance.cpp
@@ -1,6 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "angular_distance.h"
+#include "temporary_vector_store.h"
 
 using vespalib::typify_invoke;
 using vespalib::eval::TypifyCellType;
